+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- [2022-04-13] Added new Angular content
+  [\#185](https://github.com/thisdot/framework.dev/pull/185)
 - [2021-04-12] Added bottom padding to `homepageContentContainerStyle`
   [\#183](https://github.com/thisdot/framework.dev/pull/183)
   ([iaremarkus](https://github.com/iaremarkus))